This is a non-functional attempt to update ksoap2-android from commons-httpclient-3.1 to org.apache.httpcomponents-httpclient-4.x.
Apparently, a version of the Android SDK after Jorge's original patch made that switch.  This patch attempts to update the AndroidHttpTransport and AndroidServiceConnection classes to make use of the newer library.  It doesn't work, but I'm comitting this patch to save the attempt for posterity.  The problems will be reverted/resolved in the next commit.